// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/door-link-ayahs-page-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DoorLinkAyahDto } from '../models/door-link-ayah-dto';
export interface DoorLinkAyahsPageDto {
  doorId: number;
  doorVersion: number;
  isGrouped: boolean;
  items: Array<DoorLinkAyahDto>;
  linkingDataRevision: number;
  page: number;
  pageSize: number;
  totalCount: number;
  unitId: number;
}