// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/door-link-records-page-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DoorLinkRecordSummaryDto } from '../models/door-link-record-summary-dto';
export interface DoorLinkRecordsPageDto {
  doorId: number;
  doorVersion: number;
  items: Array<DoorLinkRecordSummaryDto>;
  page: number;
  pageSize: number;
  totalCount: number;
}